requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
flake8==6.1.0
greenlet==3.0.1
idna==3.4
mccabe==0.7.0
mypy==1.7.0
mypy-extensions==1.0.0
numpy==1.26.2
pandas==2.1.3
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
